Timings for busy-wait, no-busy-wait implementations
- note that no-busy-wait takes *longer*...maybe because it has to wait
  on two condition variables? i'd wager the time between consecutive
  checks of empty()/full() in a tight loop is shorter than that between
  waiting on barging_ and waiting on itemAvailable_/slotFree_...
- this effect is observed even with compiler optimization, which backs
  up the assertion that this is to do with the relative efficiency of
  the two approaches (and not some random effect, blah blah blah.)
- it would be more instructive to run each program in a tight loop,
  capturing several timings and averaging. i probably won't do that,
  though; the difference seems significant, and there isn't much to be
  learned by writing yet another for i in $(seq 1 N) one-liner :)

+++ a3/Consumer.cc
@@ -16,8 +16,8 @@ _Task Consumer {
   Consumer(uint32_t id,
            BufType& buffer,
            const int32_t delay,
-           const int32_t sentinel,
-           int32_t& sum)
+           const int64_t sentinel,
+           int64_t& sum)
     : buffer_(buffer),
       delay_(delay),
       sentinel_(sentinel),
@@ -31,14 +31,14 @@ _Task Consumer {
  private:
   BufType& buffer_;
   const int32_t delay_;
-  const int32_t sentinel_;
-  int32_t& sum_;
+  const int64_t sentinel_;
+  int64_t& sum_;
   string name_;
   void main() {
     sum_ = 0;
     while (true) {
       yield(rand() % delay_);
-      int32_t value = buffer_.remove();
+      int64_t value = buffer_.remove();
       if (value == sentinel_) {
         break;
       }
